+++ Solution/RiskOfChaos/EffectDefinitions/World/Spawn/Projectile/SpawnAirstrikes.cs
@@ -14,7 +14,7 @@ using UnityEngine.ResourceManagement.AsyncOperations;
 
 namespace RiskOfChaos.EffectDefinitions.World.Spawn.Projectile
 {
-    [ChaosEffect("spawn_airstrikes", DefaultSelectionWeight = 0.9f)]
+    [ChaosEffect("spawn_airstrikes")]
     public sealed class SpawnAirstrikes : NetworkBehaviour
     {
         static readonly SpawnUtils.NodeSelectionRules _strikePositionSelectorRules = new SpawnUtils.NodeSelectionRules(SpawnUtils.NodeGraphFlags.Ground, false, HullMask.Human, NodeFlags.None, NodeFlags.None);